# openapi/vietgara.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "VietGara API",
    "description": "Public REST/JSON API of the VietGara platform (API Specification). Generated from vietgara-proto.",
    "version": "1.1"
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "summary": "Returns the signed-in account.",
        "operationId": "AccountService_GetMe",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1Account"
            }
          },
          "default": {
            "description": "Error (API Specification Section 1.1).",
            "schema": {
              "$ref": "#/definitions/v1ErrorResponse"
            }
          }
        },
        "tags": [
          "AccountService"
        ]
      },
      "patch": {
        "summary": "Updates the signed-in account's profile and language.",
        "operationId": "AccountService_UpdateMe",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1Account"
            }
          },
          "default": {
            "description": "Error (API Specification Section 1.1).",
            "schema": {
              "$ref": "#/definitions/v1ErrorResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "account",
            "description": "Fields to change: full_name, phone, locale.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1Account"
            }
          }
        ],
        "tags": [
          "AccountService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "v1Account": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "email": {
          "type": "string",
          "readOnly": true
        },
        "phone": {
          "type": "string",
          "description": "Optional contact phone, unique across accounts; used to look staff up."
        },
        "authProvider": {
          "$ref": "#/definitions/v1AuthProvider",
          "readOnly": true
        },
        "fullName": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/v1AccountStatus",
          "readOnly": true
        },
        "emailVerified": {
          "type": "boolean",
          "description": "False until the e-mail OTP is confirmed; unverified accounts cannot use\ngarage, staff or business APIs.",
          "readOnly": true
        },
        "locale": {
          "$ref": "#/definitions/v1Locale"
        },
        "platformAdmin": {
          "type": "boolean",
          "description": "Platform operators manage subscription plans (FR-LIC-05).",
          "readOnly": true
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "readOnly": true
        }
      },
      "description": "Account is a VietGara login (owner or staff member)."
    },
    "v1AccountStatus": {
      "type": "string",
      "enum": [
        "ACCOUNT_STATUS_ACTIVE",
        "ACCOUNT_STATUS_SUSPENDED"
      ],
      "description": "AccountStatus is the lifecycle of a login account (Database Design 2.1)."
    },
    "v1AuthProvider": {
      "type": "string",
      "enum": [
        "AUTH_PROVIDER_EMAIL",
        "AUTH_PROVIDER_GOOGLE"
      ],
      "description": "AuthProvider is how the account signed up (FR-IAM-01)."
    },
...
    "v1Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "description": "Stable, machine-readable code the client translates, e.g.\nVALIDATION_ERROR, EMAIL_NOT_VERIFIED, OTP_EXPIRED, PLAN_LIMIT_REACHED."
        },
        "message": {
          "type": "string",
          "description": "English, developer-facing description; clients show a localized text\nfor `code` instead."
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1ErrorDetail"
          },
          "description": "Field-level problems for VALIDATION_ERROR."
        }
      },
      "description": "Error describes why a request failed."
    },
    "v1ErrorDetail": {
      "type": "object",
      "properties": {
        "field": {
          "type": "string",
          "description": "JSON name of the field, e.g. \"email\" or \"garage.name\"."
        },
        "issue": {
          "type": "string",
          "description": "Machine-readable problem: required, invalid, too_long, taken, ..."
        }
      },
      "description": "ErrorDetail is one invalid request field."
    },
    "v1ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/v1Error"
        }
      },
      "description": "ErrorResponse is the body of every non-2xx REST response:\n`{\"error\": {\"code\", \"message\", \"details\"}}`."
    },
...
    "v1Locale": {
      "type": "string",
      "enum": [
        "LOCALE_VI",
        "LOCALE_EN"
      ],
      "description": "Locale is the account's interface and e-mail language. Vietnamese is the\ndefault (NFR-US-01)."
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "description": "Access token issued by the auth endpoints: `Bearer \u003ctoken\u003e`.",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}